Migrate ComponentRegistry
