+++ packages/fabrix/src/renderers/form.tsx
@@ -12,6 +12,7 @@ import {
 } from "./shared";
 import { buildAjvSchema } from "./form/validation";
 import { ajvResolver } from "./form/ajvResolver";
+import { printIntrospectionSchema } from "graphql";
 
 export type FormFields = FieldConfigByType<"form">["configs"]["fields"];
 export type FormField = FormFields[number];
@@ -70,12 +71,13 @@ export const FormRenderer = ({
     return <Loader />;
   }
 
-  const component = context.componentRegistry.components.default?.form;
+  const component = context.componentRegistry.getDefaultComponentByType("form");
   if (!component) {
     return;
   }
 
   return createElement(component, {
+    name: rootField.name,
     renderFields: () => {
       return <FormProvider {...formContext}>{renderFields()}</FormProvider>;
     },
@@ -87,6 +89,7 @@ export const FormRenderer = ({
     renderReset: (resetRenderer) =>
       resetRenderer({ reset: () => formContext.reset() }),
     className: `fabrix form col-row ${className ?? ""}`,
+    customProps: {},
   });
 };
 
@@ -103,11 +106,10 @@ const renderField = (props: {
   }
 
   const component = fieldConfig.componentType?.name
-    ? context.componentRegistry.getCustom(
+    ? context.componentRegistry.getUnitComponentByName<"formField">(
         fieldConfig.componentType.name,
-        "formField",
       )
-    : context.componentRegistry.components.default?.formField;
+    : context.componentRegistry.getDefaultComponentByType("formField");
   if (!component) {
     return;
   }
